axi_dmac: external sync support
With external synchronization external components to the DMAC  will be able
to throttle the consumption of descriptors queued by the software.

A transfer will start only after the assertion of the external sync
signal. The sync signal can be either in source or destination clock domain or
both.
This feature does not ensures fixed latency from the assertion of
external sync signal and the availability of the data at the destination
interface.