axi_dmac: simple framelock; wait for master selectable
This change adds support for 'Simple Framelock' mode where the master (writer) DMAC
will not look after the currently read buffer by the slave DMAC, it will
cycle through the buffers regardless of the reader. The slave will always read
a buffer at a predefined distance from the one currently accessed by the master.
This is useful where the writer and reader frame rate match and a predefined
fixed number of frame delay must be produced between reader and writer.

In Simple Framelock mode if the 'wait for master' is enabled the reader will output
a frame only after the master wrote one to the memory. The slave will
follow the master at a defined fixed distance.
If the 'wait for master' is not enabled the slave will start reading a
buffer whenever it completed a previous buffer and receives an external sync
signal if the external synchronization support is enabled.